Fix non-increasing ramp.fits LUT values
Pete Ratzlaff wrote to the pdl-general list on 2023-07-05 about
a problem with the PDL::Graphics::LUT default ramp having values
that are not always increasing--there is a wrinkle in the LUT.

This commit fixes that for both the ramp and the rgb table, and also
aligns the values of the neg.fits ramp so that they agree with
the new values in ramp.fits.  xlinvals was used to make the new
ramp and RGB LUT values.